Extract durable control-turn collaborators
