+++ tests/test_hook_sender.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import json
-from dataclasses import replace
 from typing import TYPE_CHECKING
 from unittest.mock import AsyncMock, MagicMock, patch
 
@@ -1169,7 +1168,7 @@ async def test_dispatch_text_message_runs_message_received_before_command_parsin
     bot._conversation_resolver.extract_dispatch_context = AsyncMock(
         return_value=dispatch_context_result(_dispatch_context(bot)),
     )
-    bot._turn_controller._execute_command = AsyncMock()
+    bot._command_turn_executor.execute_if_owned = AsyncMock(return_value=True)
     turn_store = unwrap_extracted_collaborator(bot._turn_store)
     turn_store.record_turn = MagicMock()
 
@@ -1179,7 +1178,7 @@ async def test_dispatch_text_message_runs_message_received_before_command_parsin
     )
 
     assert hook_calls == ["called"]
-    bot._turn_controller._execute_command.assert_not_awaited()
+    bot._command_turn_executor.execute_if_owned.assert_not_awaited()
     turn_store.record_turn.assert_not_called()
 
 
@@ -1211,10 +1210,7 @@ async def test_prepare_dispatch_compacts_all_source_events_when_hooks_suppress_b
     turn_store = unwrap_extracted_collaborator(bot._turn_store)
     turn_store.record_turn = MagicMock()
     settle_ignored = AsyncMock()
-    bot._turn_controller.deps = replace(
-        bot._turn_controller.deps,
-        settle_ignored_dispatch_sources=settle_ignored,
-    )
+    replace_turn_controller_deps(bot, settle_ignored_sources=settle_ignored)
 
     dispatch = await bot._turn_controller._prepare_dispatch(
         room,
